+++ assets/index-react.ejs
@@ -10,7 +10,7 @@
   <noscript>
     <p>This application requires JavaScript.</p>
   </noscript>
-  <div class="h100" id="react-root"></div>
+  <div id="react-root"></div>
 </body>
 
 </html>
