+++ src/react-demo/Logo.tsx
@@ -1,0 +1,44 @@
+import React from "react";
+
+export const Logo = () => (
+  <svg
+    width="20"
+    height="20"
+    viewBox="0 0 20 20"
+    fill="none"
+    xmlns="http://www.w3.org/2000/svg"
+  >
+    <path
+      d="M19.0616 14.5308L16.1901 11.6592C14.0524 12.2301 11.677 11.677 10 9.99998C8.32296 8.32294 7.76987 5.94757 8.34073 3.80991L5.46918 0.938362C4.21803 -0.312788 2.18952 -0.312787 0.938366 0.938362C-0.312784 2.18951 -0.312783 4.21803 0.938366 5.46917L5.4692 10L0.938364 14.5308C-0.312787 15.782 -0.312788 17.8105 0.938362 19.0616C2.18951 20.3128 4.21803 20.3128 5.46918 19.0616L10 14.5308L14.5308 19.0616C15.782 20.3128 17.8105 20.3128 19.0616 19.0616C20.3128 17.8105 20.3128 15.7819 19.0616 14.5308Z"
+      fill="url(#paint0_linear_1638_3672)"
+    />
+    <path
+      d="M12.2654 3.20378C11.0143 4.45492 11.0143 6.48344 12.2654 7.73459C13.5166 8.98574 15.5451 8.98574 16.7962 7.73459C18.0474 6.48344 19.7024 0.297616 19.7024 0.297616C19.7024 0.297616 13.5166 1.95263 12.2654 3.20378Z"
+      fill="url(#paint1_linear_1638_3672)"
+    />
+    <defs>
+      <linearGradient
+        id="paint0_linear_1638_3672"
+        x1="0"
+        y1="20"
+        x2="20"
+        y2="-2.38419e-06"
+        gradientUnits="userSpaceOnUse"
+      >
+        <stop stop-color="#886DFF" />
+        <stop offset="1" stop-color="#FF97F9" />
+      </linearGradient>
+      <linearGradient
+        id="paint1_linear_1638_3672"
+        x1="0"
+        y1="20"
+        x2="20"
+        y2="-2.38419e-06"
+        gradientUnits="userSpaceOnUse"
+      >
+        <stop stop-color="#886DFF" />
+        <stop offset="1" stop-color="#FF97F9" />
+      </linearGradient>
+    </defs>
+  </svg>
+);
