+++ src/react-demo/icons/History.tsx
@@ -1,0 +1,18 @@
+import React from "react";
+
+export const History = () => (
+  <svg
+    width="16"
+    height="16"
+    viewBox="0 0 16 16"
+    fill="none"
+    xmlns="http://www.w3.org/2000/svg"
+  >
+    <path
+      fill-rule="evenodd"
+      clip-rule="evenodd"
+      d="M1.64288 3.14365L0.42697 1.92781C0.392008 1.89276 0.347425 1.86888 0.298874 1.8592C0.250324 1.84952 0.199991 1.85447 0.154258 1.87342C0.108524 1.89238 0.0694496 1.92449 0.0419873 1.96568C0.014525 2.00687 -8.81759e-05 2.05528 4.00297e-07 2.10479V5.75032C4.00297e-07 5.88831 0.111992 6.00029 0.249982 6.00029H3.89572C3.94523 6.00038 3.99365 5.98577 4.03484 5.95831C4.07603 5.93085 4.10814 5.89178 4.1271 5.84604C4.14605 5.80031 4.15101 5.74998 4.14132 5.70144C4.13164 5.65289 4.10776 5.60831 4.07271 5.57335L2.71481 4.21552C3.59998 2.97932 4.89056 2.09257 6.362 1.70954C7.83345 1.32652 9.39262 1.47146 10.7682 2.11916C12.1438 2.76686 13.2488 3.87631 13.8909 5.25449C14.533 6.63267 14.6716 8.19233 14.2825 9.66212C13.8935 11.1319 13.0014 12.4188 11.7615 13.2988C10.5217 14.1789 9.01247 14.5964 7.49655 14.4788C5.98062 14.3612 4.55391 13.7158 3.46465 12.655C2.3754 11.5942 1.69254 10.1852 1.53489 8.67296C1.51447 8.475 1.41624 8.29326 1.26182 8.16772C1.10739 8.04218 0.909419 7.98312 0.711449 8.00354C0.51348 8.02396 0.33173 8.12218 0.206183 8.2766C0.0806366 8.43102 0.0215772 8.62898 0.0419974 8.82694C0.234372 10.6779 1.06608 12.404 2.39393 13.7079C3.72179 15.0118 5.46271 15.8121 7.31701 15.9708C9.1713 16.1296 11.0229 15.637 12.5531 14.5778C14.0834 13.5186 15.1964 11.959 15.7007 10.1677C16.205 8.37633 16.0689 6.46522 15.316 4.76336C14.563 3.06149 13.2403 1.67536 11.5754 0.843575C9.9106 0.0117949 8.00785 -0.213586 6.19475 0.206231C4.38165 0.626048 2.77164 1.6648 1.64188 3.14365H1.64288ZM7.74945 4.00055C7.94834 4.00055 8.13909 4.07955 8.27974 4.22019C8.42038 4.36082 8.49939 4.55156 8.49939 4.75045V7.74207L10.5272 8.55397C10.7059 8.63187 10.8473 8.77628 10.9213 8.95661C10.9954 9.13694 10.9963 9.33901 10.9238 9.51999C10.8514 9.70097 10.7114 9.84664 10.5334 9.92613C10.3554 10.0056 10.1534 10.0127 9.97029 9.9458L7.47047 8.94592C7.33144 8.89021 7.21229 8.79415 7.12835 8.67012C7.04441 8.54609 6.99953 8.39977 6.9995 8.25001V4.75045C6.9995 4.55156 7.07851 4.36082 7.21915 4.22019C7.3598 4.07955 7.55055 4.00055 7.74945 4.00055Z"
+      fill="#5F5F6C"
+    />
+  </svg>
+);
